web/requirements.txt
websockets==11.0.3
python-chess==1.999
